+++ lib/minipack/helper.rb
@@ -9,7 +9,7 @@ module Minipack::Helper
   #   <%= asset_bundle_path 'icon/favicon.ico' %> # => "/assets/web/pack/icon/favicon-1016838bab065ae1e122.ico"
   def asset_bundle_path(name, manifest: nil, **options)
     manifest = get_manifest_by_key(manifest)
-    asset_path(manifest.lookup!(name.to_s), **options)
+    asset_path(manifest.lookup!(name.to_s).path, **options)
   end
 
   # Example:
@@ -20,7 +20,9 @@ module Minipack::Helper
   #   <%= javascript_bundle_tag 'orders/app'  %> # =>
   #   <script src="/assets/web/pack/orders/app-1016838bab065ae1e314.js"></script>
   def javascript_bundle_tag(*names, manifest: nil, **options)
-    javascript_include_tag(*sources_from_manifest(names, 'js', key: manifest), **options)
+    entries_from_manifest(names, 'js', key: manifest).map { |entry|
+      javascript_include_tag(entry.path, **options_for(entry, options))
+    }.join("\n").html_safe
   end
 
   # Creates script tags that references the js chunks from entrypoints when using split chunks API.
@@ -53,7 +55,9 @@ module Minipack::Helper
   #    href="/assets/web/pack/orders/style-1016838bab065ae1e122.css" />
   def stylesheet_bundle_tag(*names, manifest: nil, **options)
     if Minipack.configuration.extract_css?
-      stylesheet_link_tag(*sources_from_manifest(names, 'css', key: manifest), **options)
+      entries_from_manifest(names, 'css', key: manifest).map { |entry|
+        stylesheet_link_tag(entry.path, **options_for(entry, options))
+      }.join("\n").html_safe
     end
   end
 
@@ -85,23 +89,30 @@ module Minipack::Helper
   #   <img src="/assets/pack/icon-1016838bab065ae1e314.png" width="16" height="10" alt="Edit Entry" />
   def image_bundle_tag(name, manifest: nil, **options)
     manifest = get_manifest_by_key(manifest)
-    image_tag(manifest.lookup!(name.to_s), **options)
+    image_tag(manifest.lookup!(name.to_s).path, **options)
   end
 
   private
 
-  def sources_from_manifest(names, ext, key: nil)
+  def entries_from_manifest(names, ext, key: nil)
     manifest = get_manifest_by_key(key)
     names.map { |name| manifest.lookup!(name.to_s + '.' + ext) }
   end
 
   def sources_from_manifest_entrypoints(names, type, key: nil)
     manifest = get_manifest_by_key(key)
-    names.map { |name| manifest.lookup_pack_with_chunks!(name, type: type) }.flatten.uniq
+    names.map { |name| manifest.lookup_pack_with_chunks!(name, type: type).entries.map(&:path) }.flatten.uniq
   end
 
   def get_manifest_by_key(key = nil)
     repository = Minipack.configuration.manifests
     key.nil? ? repository.default : repository.get(key)
+  end
+
+  # @param [Minipack::Manifest::Entry] entry
+  # @param [Hash] options
+  def options_for(entry, options)
+    return options unless entry.integrity
+    options.merge(integrity: entry.integrity)
   end
 end